fix: correction to toAdjacencyMap graph method
